Cherry Pick - RPi builds no longer need `no-interactive` workaround (#20600)
* Switch from readline to editline (#20330)

Readline is GPL'd software and may not be available in the build
environment.

Vendor editline which is permissively licensed and drop it in as a
replacement for readline.

* Don't sweat interactive mode in build_examples.py (#20389)

As of 3340fc2b62c ("Switch from readline to editline (#20330)") there is
no longer an issue with respect to the availability of libreadline.

* Target non-interactive Linux example build in CI (#20635)

Co-authored-by: Michael Spang <spang@google.com>